Migrate create_handler to __init__.py
Avoid circular import in env_handler
Lazy import so we don't load unused handlers
