+++ mbrl/util/env_handler.py
@@ -99,29 +99,6 @@ def _legacy_make_env(
     return env, term_fn, reward_fn
 
 
-def create_handler(env_name: str):
-    """Creates a new environment handler from its string description.
-
-    Args:
-        env_name (str): the string description of the environment. Valid options are:
-
-          - "dmcontrol___<domain>--<task>": a Deep-Mind Control suite environment
-            with the indicated domain and task (e.g., "dmcontrol___cheetah--run".
-          - "gym___<env_name>": a Gym environment (e.g., "gym___HalfCheetah-v2").
-          - "cartpole_continuous": a continuous version of gym's Cartpole environment.
-          - "pets_halfcheetah": the implementation of HalfCheetah used in Chua et al.,
-            PETS paper.
-          - "ant_truncated_obs": the implementation of Ant environment used in Janner et al.,
-            MBPO paper.
-          - "humanoid_truncated_obs": the implementation of Humanoid environment used in
-            Janner et al., MBPO paper.
-
-    Returns:
-        (EnvHandler): A handler for the associated gym environment
-    """
-    pass
-
-
 class Freeze(ABC):
     """ Abstract base class for freezing various gym backends """
